Electronica modificacio preu

+++ Practica SapaMercat/src/Electronica.java
@@ -6,9 +6,8 @@ public class Electronica extends Producte{
         this.diesDeGarantia = diesDeGarantia;
     }
 
-    public double getPreu() {
-        //
-        return preu + preu * ((double) diesDeGarantia / 365) * 0.1;
+    public float getPreu() {
+        return (float) (preu + preu * ( diesDeGarantia / 365) * 0.1);
     }
 
     @Override
